improve: syntax, variable and field names
This commit improves overall syntax of the project, following better PerformanC standards.

+++ external/tablec.c
@@ -114,6 +114,6 @@ struct tablec_bucket *tablec_get(struct tablec_ht *tablec, char *key) {
   return NULL;
 }
 
-size_t tablec_full(struct tablec_ht *tablec) {
-  return tablec->capacity == tablec->length ? -1 : tablec->capacity - tablec->length;
+int tablec_full(struct tablec_ht *tablec) {
+  return (int)(tablec->capacity == tablec->length) ? -1 : (int)(tablec->capacity - tablec->length);
 }
