+++ external/tablec.h
@@ -28,6 +28,6 @@ void tablec_del(struct tablec_ht *tablec, char *key);
 
 struct tablec_bucket *tablec_get(struct tablec_ht *tablec, char *key);
 
-size_t tablec_full(struct tablec_ht *tablec);
+int tablec_full(struct tablec_ht *tablec);
 
 #endif /* TABLEC_H */
